Adds multiple choice component
Adds a new component that allows users to select a single option from a list of choices.

This component is implemented using Angular Material radio buttons and supports disabling the component and setting a default value. It also implements the ControlValueAccessor interface, allowing it to be used with Angular forms.